[22-10-06] 스케줄러 삭제
