+++ src/main/java/com/havit/finalbe/controller/MainController.java
@@ -2,7 +2,6 @@ package com.havit.finalbe.controller;
 
 import com.havit.finalbe.dto.response.AllGroupListResponseDto;
 import com.havit.finalbe.dto.response.UnchallengerResponseDto;
-import com.havit.finalbe.scheduler.MainScheduler;
 import com.havit.finalbe.security.userDetail.UserDetailsImpl;
 import com.havit.finalbe.service.MainService;
 import io.swagger.annotations.Api;
@@ -23,7 +22,6 @@ import java.util.List;
 public class MainController {
 
     private final MainService mainService;
-    private final MainScheduler mainScheduler;
 
     @Operation(summary = "그룹 검색", description = "그룹 제목, 그룹 작성자 닉네임으로 전체 그룹을 검색합니다.")
     @GetMapping("/search")
@@ -35,6 +33,6 @@ public class MainController {
     @Operation(summary = "알림 보낼 그룹원 목록 조회", description = "매일 챌린지에 참여하지 않은 멤버 목록을 불러옵니다.")
     @GetMapping("/")
     public List<UnchallengerResponseDto> searchGroup(@AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return mainScheduler.getUnchallengerList(userDetails);
+        return mainService.getUnchallengerList(userDetails);
     }
 }
